refactor flat management

+++ app/Http/Controllers/ServiceManagement/ServiceController.php
@@ -2,9 +2,14 @@
 
 namespace App\Http\Controllers\ServiceManagement;
 
+use App\Enum\AbilityEnum;
 use App\Http\Controllers\Controller;
 use App\Http\Requests\ServiceManagement\ServiceRequest;
 use App\Models\ServiceManagement\Service;
+use Illuminate\Auth\Access\AuthorizationException;
+use Illuminate\Contracts\Foundation\Application;
+use Illuminate\Contracts\View\Factory;
+use Illuminate\Contracts\View\View;
 use function __;
 use function redirect;
 use function view;
@@ -16,7 +21,10 @@ class ServiceController extends Controller
         $this->middleware('auth');
     }
 
-    public function index()
+    /**
+     * @throws AuthorizationException
+     */
+    public function index(): Factory|View|Application
     {
         $this->authorize('view', Service::class);
         $records = Service::orderBy('name', 'ASC')->get();
@@ -24,44 +32,41 @@ class ServiceController extends Controller
             'pageTitle' => __('general.service'),
             'records' => $records,
         ];
-        return view('dashboard.definition.service.index',$params);
+        return view('dashboard.service-management.services.index',$params);
     }
 
-    public function create()
+    /**
+     * @throws AuthorizationException
+     */
+    public function create(): Factory|View|Application
     {
-        // $this->authorize(\App\Enum\AbilityEnum::CREATE, Service::class);
+       $this->authorize(AbilityEnum::CREATE, Service::class);
         $params = [
             'pageTitle' => __('general.new_service'),
         ];
-        return view('dashboard.definition.service.create',compact('params'));
+        return view('dashboard.service-management.services.create',compact('params'));
     }
 
 
+    /**
+     * @throws AuthorizationException
+     */
     public function store(ServiceRequest $request)
     {
-        // $this->authorize('create', Service::class);
+         $this->authorize('create', Service::class);
         if ($request->createData()) {
-            if ($request->saveNew) {
-                return redirect()->route('dashboard.service.create')
-                    ->with('success', __('general.record_created_successfully'));
-            } else {
-                return redirect()->route('dashboard.services.index')
-                    ->with('success', __('general.record_created_successfully'));
-            }
+            return redirect()->route('dashboard.services.index')
+                ->with('success', __('general.record_created_successfully'));
         }
     }
 
 
-    public function show($id)
+    /**
+     * @throws AuthorizationException
+     */
+    public function edit($id): Factory|View|Application
     {
-        $this->authorize('view', Service::class);
-        //
-    }
-
-
-    public function edit($id)
-    {
-        // $this->authorize('update', Service::class);
+         $this->authorize('update', Service::class);
         $model = Service::findorFail($id);
 
         $params = [
@@ -69,22 +74,28 @@ class ServiceController extends Controller
             'model' => $model,
         ];
 
-        return view('dashboard.definition.service.edit', $params);
+        return view('dashboard.service-management.services.edit', $params);
     }
 
 
+    /**
+     * @throws AuthorizationException
+     */
     public function update(ServiceRequest $request, $id)
     {
-        // $this->authorize('update', Service::class);
+         $this->authorize('update', Service::class);
         if ($request->updateData($id)) {
             return redirect()->route('dashboard.services.index')
                 ->with('success', __('general.record_updated_successfully'));
         }
     }
 
-    public function destroy(ServiceRequest $request,$id)
+    /**
+     * @throws AuthorizationException
+     */
+    public function destroy(ServiceRequest $request, $id)
     {
-        // $this->authorize('delete', Service::class);
+         $this->authorize('delete', Service::class);
         if ($request->deleteData($id)) {
             return redirect()->route('dashboard.services.index')
                 ->with('success', __('general.record_deleted_successfully'));
